+++ Xure.Api/Services/Products/ProductRepository.cs
@@ -12,7 +12,7 @@ namespace Xure.Api.Services
         {
 
         }
-        public Product GetById(int id)
+        public Product GetById(long id)
         {
             return GetWithInclude(c => c.Id == id, c => c.Price.PriceHistory, c => c.Category, c => c.Brands, c=> c.ProductSpecificationsValues ).FirstOrDefault(x => x.Id == id);
         }
